rm front, mk subject api, cronjobs fixed
